+++ src/lib.rs
@@ -15,10 +15,18 @@ pub mod ffi;
 pub mod hid;
 
 pub use error::HidError;
-pub use hid::{HidDevice, HidDeviceInfo, HidElement, HidManager, HidUsage, ReportSubscription};
+pub use hid::{
+    DeviceMatch, ElementMatch, HidCollectionType, HidDevice, HidDeviceInfo, HidElement,
+    HidElementType, HidInputReport, HidManager, HidReportType, HidUsage, HidValue, HidValueScale,
+    ReportSubscription, TimestampedReportSubscription, ValueSubscription,
+};
 
 /// Common imports.
 pub mod prelude {
     pub use crate::error::HidError;
-    pub use crate::hid::{HidDevice, HidDeviceInfo, HidElement, HidManager, HidUsage, ReportSubscription};
+    pub use crate::hid::{
+        DeviceMatch, ElementMatch, HidCollectionType, HidDevice, HidDeviceInfo, HidElement,
+        HidElementType, HidInputReport, HidManager, HidReportType, HidUsage, HidValue,
+        HidValueScale, ReportSubscription, TimestampedReportSubscription, ValueSubscription,
+    };
 }